MAINT: Separate publish app into components
